equation to edit text (need to put back)

+++ src/graph.cpp
@@ -233,10 +233,10 @@ void gengraph(state* s, int index) {
 			s->graphs[index]->verticies[i + 1] /= (s->set.ymax - s->set.ymin) / 20;
 			s->graphs[index]->verticies[i + 2] /= (s->graphs[index]->zmax - s->graphs[index]->zmin) / 20;
 		}
-		s->graphs[index]->xmin = s->set.xmin;
-		s->graphs[index]->xmax = s->set.xmax;
-		s->graphs[index]->ymin = s->set.ymin;
-		s->graphs[index]->ymax = s->set.ymax;
+		s->graphs[index]->xmin = (float)s->set.xmin;
+		s->graphs[index]->xmax = (float)s->set.xmax;
+		s->graphs[index]->ymin = (float)s->set.ymin;
+		s->graphs[index]->ymax = (float)s->set.ymax;
 		s->graphs[index]->zmin = -10;
 		s->graphs[index]->zmax = 10;
 	}
